### 2021-11-19 #3935 @whedon
[PRE REVIEW]: pyRS: A Python package for the reduction and analysis of neutron residual stress data
**Submitting author:** @fanchercm (<a href="http://orcid.org/0000-0002-3952-5168">Chris Fancher</a>)
**Repository:** <a href="https://github.com/neutrons/PyRS/" target ="_blank">https://github.com/neutrons/PyRS/</a>
**Version:** v1.0.1
