### 2021-11-22 #3935 @rkurchin
I can do it, but I'm hoping not to take on any new stuff this week so I can have a bit of a break over the holiday. Is it okay if I pick this up in a week's time?
