### 2021-11-23 #3935 @rkurchin
No problem, go for it!

—————————————
Rachel C. Kurchin, PhD
Postdoctoral Researcher (formerly MolSSI and MFI Fellow)
Carnegie Mellon Mechanical Engineering and Materials Science
https://rkurchin.github.io
