+++ .travis.yml
@@ -1,74 +1,0 @@
-language: python
-dist: trusty
-
-services:
-  - xvfb
-
-cache:
-  directories:
-    - $HOME/.cache/pip
-
-matrix:
-  include:
-  - env: CONDA=3.6
-
-before_install:
-- |
-  # Install conda
-  "/sbin/start-stop-daemon --start --quiet --pidfile /tmp/custom_xvfb_99.pid --make-pidfile --background --exec /usr/bin/Xvfb -- :99 -ac -screen 0 1280x1024x16"
-  if [ "$TRAVIS_OS_NAME" = "linux" ]; then export CXX=g++-4.8 CC=gcc-4.8; fi
-  if [ "$TRAVIS_OS_NAME" = "linux" ]; then OS=Linux-x86_64; else OS=MacOSX-x86_64; fi
-  wget -O miniconda.sh https://repo.continuum.io/miniconda/Miniconda${CONDA:0:1}-latest-$OS.sh
-  bash miniconda.sh -b -p $HOME/miniconda
-  export PATH="$HOME/miniconda/bin:$PATH"
-  conda init bash
-
-install:
-    # Setup conda activate/deactivate commands
-  - conda init bash
-  - source $(conda info --root)/etc/profile.d/conda.sh
-  - conda info --root
-
-    # Conda config - behavior and channel setup
-  - conda config --set always_yes yes --set changeps1 no --set anaconda_upload no
-  - conda config --add channels conda-forge --add channels mantid --add channels mantid/label/nightly
-
-script:
-    # TEST: Create test environment and test build
-  - conda install mamba
-  - mamba create -q -n PyRS_test python=$CONDA mantid-workbench>=4.2 --file requirements.txt --file requirements_dev.txt
-  - conda info --envs
-  - conda activate PyRS_test
-  - python --version
-
-  - |
-    # Mantid pre-requiste - create a properties file that turns off network access
-    mkdir ~/.mantid
-    echo "CheckMantidVersion.OnStartup=0" > ~/.mantid/Mantid.user.properties
-    echo "UpdateInstrumentDefinitions.OnStartup=0" >> ~/.mantid/Mantid.user.properties
-    echo "usagereports.enabled=0" >> ~/.mantid/Mantid.user.properties
-    export DISPLAY=:99.0
-    sleep 3
-
-  # lint the code and generate an error if a warning is introduced
-  - flake8 . --count
-  # static analysis for typing
-  - yes | pip install types-pkg_resources types-six types-toml
-  - mypy pyrs scripts tests
-
-  # run some basic import tests to confirm things are working
-  - python -c "import mantid"
-  - python -c "import qtpy"
-  - python -c "import mantidqt"
-
-  # mantid workbench tests
-  - xvfb-run --server-args="-screen 0 640x480x24" --auto-servernum mantidworkbench --help
-  - echo "import time;time.sleep(5)" > workbenchshutdown.py
-  # this requires the mantid-workbench package to be fixed
-  # - xvfb-run --server-args="-screen 0 640x480x24" --auto-servernum mantidworkbench -q -x workbenchshutdown.py
-
-  # run PyRS tests
-  - xvfb-run --server-args="-screen 0 640x480x24" --auto-servernum pytest --cov=pyrs tests --disable-pytest-warnings -vv -r w
-
-after_success:
-  - codecov
