### 2021-12-20 2f770a8
Added option to reduce data from a mantid workspace
