+++ pyrs/core/live_data_conversion.py
@@ -1,0 +1,594 @@
+"""
+Convert HB2B NeXus file to Hidra project file for further reduction
+"""
+from mantid.kernel import Logger, BoolTimeSeriesProperty, FloatFilteredTimeSeriesProperty, FloatTimeSeriesProperty
+from mantid.kernel import Int32TimeSeriesProperty, Int64TimeSeriesProperty, Int32FilteredTimeSeriesProperty,\
+    Int64FilteredTimeSeriesProperty
+from mantid.simpleapi import mtd, DeleteWorkspace, LoadMask, RemoveLogs, CopyLogs, CreateSampleWorkspace
+import numpy as np
+import os
+from pyrs.core import workspaces
+from pyrs.core.instrument_geometry import DENEXDetectorGeometry, HidraSetup
+from pyrs.core import MonoSetting  # type: ignore
+from pyrs.dataobjects import HidraConstants  # type: ignore
+from pyrs.projectfile import HidraProjectFile, HidraProjectFileMode  # type: ignore
+from pyrs.utilities import checkdatatypes
+
+SUBRUN_LOGNAME = 'scan_index'
+NUM_PIXEL_1D = 1024
+HIDRA_PIXEL_NUMBER = NUM_PIXEL_1D * NUM_PIXEL_1D
+PIXEL_SIZE = 0.3 / NUM_PIXEL_1D
+ARM_LENGTH = 0.985
+
+DEFAULT_KEEP_LOGS = ['experiment_identifier', 'run_number', 'run_title', 'file_notes', 'start_time', 'end_time',
+                     'SampleId', 'SampleName', 'SampleDescription', 'StrainDirection', 'hklPhase', 'Wavelength',
+                     'Filename', 'sub-run', 'duration', 'mrot', 'mtilt', 'mb220', 'mb511', 'ISD', 'ISR:X:Gap',
+                     'ISR:Y:Gap', 'DOX', 'DOY', 'DOR', 'omega', '2theta', 'phi', 'chi', 'sx', 'sy', 'sz',
+                     'vx', 'vy', 'vz', 'omegaSetpoint', '2thetaSetpoint', 'phiSetpoint', 'chiSetpoint', 'sxSetpoint',
+                     'sySetpoint', 'szSetpoint', 'scan_index', 'duration']
+
+SWEEPING_LOGS = ['HB2B:CS:Sweep:Control', 'HB2B:CS:Sweep:Device']
+
+
+def convert_pulses_to_datetime64(h5obj):
+    '''The h5object is the h5py handle to ``event_time_zero``. This only supports pulsetimes in seconds'''
+    if h5obj.attrs['units'].decode() != 'second':
+        raise RuntimeError('Do not understand time units "{}"'.format(h5obj.attrs['units']))
+
+    # the value is number of seconds as a float
+    pulse_time = h5obj.value
+
+    # Convert deltas to times with units. This has to be done through
+    # nanoseconds because numpy truncates things to integers during the conversion
+    pulse_time = pulse_time * 1.e9 * np.timedelta64(1, 'ns')
+
+    # get absolute offset and convert to absolute time
+    start_time = np.datetime64(h5obj.attrs['offset'])
+
+    return pulse_time + start_time
+
+
+def calculate_sub_run_time_average(log_property, time_filter) -> float:
+    '''Determine the time average value of the supplied log'''
+    if log_property.size() == 1:  # single value property just copy
+        time_average_value = log_property.value
+    elif time_filter is None:  # no filtering means use all values
+        time_average_value = log_property.timeAverageValue()
+    else:
+        # filter and get time average value
+        if isinstance(log_property, FloatTimeSeriesProperty):
+            filtered_tsp = FloatFilteredTimeSeriesProperty(log_property, time_filter)
[…]
