+++ .github/workflows/ci.yml
@@ -6,7 +6,7 @@ jobs:
 
   ci:
     runs-on: ubuntu-latest
-    timeout-minutes: 10
+    timeout-minutes: 100
 
     defaults:
       run:
