+++ .github/workflows/ci.yml
@@ -1,6 +1,6 @@
 name: PyRS CI
 
-on: [push, pull_request]
+on: [pull_request]
 
 jobs:
 
