### 2021-12-21 26b35e7
Changed test_manual_reduction to skip on github actions
