### 2021-12-21 2e5f70c
Changed test_manual_reduction to skip on github actions
