### 2021-12-21 15f1d6d
Changed test_manual_reduction to skip on github actions
