### 2021-12-21 b272fdd
Changed test_manual_reduction to skip on github actions
