+++ tests/ui/test_pyrslauncher.py
@@ -1,9 +1,12 @@
 from pyrs.interface.pyrs_main import PyRSLauncher
 from qtpy import QtCore
+from tests.conftest import ON_GITHUB_ACTIONS  # set to True when running on build servers
+import pytest
 
 wait = 100
 
 
+@pytest.mark.skipif(ON_GITHUB_ACTIONS, reason='Test hangs on github CI')
 def test_launcher(qtbot):
     main_window = PyRSLauncher()
     qtbot.addWidget(main_window)
