### 2021-12-21 17f5db2
added CI tests on push and pull
