+++ .github/workflows/ci.yml
@@ -1,6 +1,6 @@
 name: PyRS CI
 
-on: [pull_request]
+on: [push, pull_request]
 
 jobs:
 
