### 2022-01-28 #4104 @whedon
[REVIEW]: pyRS: A Python package for the reduction and analysis of neutron residual stress data
**Submitting author:** @fanchercm (<a href="http://orcid.org/0000-0002-3952-5168">Chris Fancher</a>)
**Repository:** https://github.com/neutrons/PyRS/
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.1

- [x] #4105
