### 2022-02-03 neutrons/PyRS#778 @andyfaff
JOSS: PyRS is not installable according to the supplied instructions on macOS and on Linux
I am reviewing the JOSS manuscript in https://github.com/openjournals/joss-reviews/issues/4104, and am trying to install pyrs.

The setup.py file indicates that the package is OS independent (`"Operating System :: OS Independent",`). I am trying to install the software on macOS. However, when I try to install using the instructions in [README.rst](https://github.com/neutrons/PyRS/blob/next/README.rst) the install fails (as of commit 73c1e30275ce56a1e75207c26875ad2b2be8478c). Typing

`conda create -n pyrs -c mantid -c mantid/label/nightly mantid-workbench -c conda-forge  --file requirements.txt --file requirements_dev.txt`

gives the following error:

```
Collecting package metadata (current_repodata.json): done
Solving environment: failed with repodata from current_repodata.json, will retry with next repodata source.
Solving environment: failed

PackagesNotFoundError: The following packages are not available from current channels:

  - mantid-workbench[version='<=5.2']

Current channels:

  - https://conda.anaconda.org/mantid/osx-64
  - https://conda.anaconda.org/mantid/noarch
  - https://conda.anaconda.org/mantid/label/nightly/osx-64
  - https://conda.anaconda.org/mantid/label/nightly/noarch
  - https://conda.anaconda.org/conda-forge/osx-64
  - https://conda.anaconda.org/conda-forge/noarch
  - https://repo.anaconda.com/pkgs/main/osx-64
  - https://repo.anaconda.com/pkgs/main/noarch
  - https://repo.anaconda.com/pkgs/r/osx-64
  - https://repo.anaconda.com/pkgs/r/noarch

To search for alternate channels that may provide the conda package you're
looking for, navigate to

    https://anaconda.org

and use the search bar at the top of the page.
```

When I visit the [mantid/packages/mantid-workbench conda channel page](https://anaconda.org/mantid/mantid-workbench/files) there are only Linux files available. This is at odds with the `OS Independent` attribute in the `setup.py` file.

I then went to the effort of making an Ubuntu virtual machine and tried to install that package, but did not succeed as the conda package solve did not complete (presumably for the Mantid workbench dependency).
