### 2022-02-03 neutrons/PyRS#779 @andyfaff
JOSS: it's not possible to verify the functional claims of this software
I'm trying to review the [JOSS paper submission](https://github.com/openjournals/joss-reviews/issues/4104). It's not currently possible to verify the functional claims of the software because the PyRS package is not currently installable (as of 73c1e30275ce56a1e75207c26875ad2b2be8478c) from the instructions in [README.rst](https://github.com/neutrons/PyRS/blob/next/README.rst). #778
