### 2022-02-03 neutrons/PyRS#780 @andyfaff
JOSS: please provide examples of how to use the software (ideally to solve real-world analysis problems).
The JOSS instructions ask ["do the authors include examples of how to use the software (ideally to solve real-world analysis problems)."](https://github.com/openjournals/joss-reviews/issues/4104)

A brief overview of the types of processing enabled by the package (data reduction/analysis/autoreduction) is given in the paper and on readthedocs, but there are no clear cut examples given in either of those two sources.
