### 2022-02-03 neutrons/PyRS#781 @andyfaff
JOSS: Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?
The modules covered in the API documentation (on readthedocs) are satisfactory. However, there are many other functions/objects that are not documented clearly. With Python there is an expectation that objects/functions not prefixed by `_` are public. I would presume this JOSS guideline is asking that those public functions be documented clearly. Take, for example, `pyrs.core.instrument_geometry.HidraSetup`, this class requires a `detector_setup` parameter for instantiation, but it's not specified what that parameter is (float/int/object/etc).
