### 2022-02-03 neutrons/PyRS#782 @andyfaff
JOSS: there are no clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support
There are no instructions for any of these three points in the online readthedocs documentation, or in the code repository itself. However, one could argue that because the project is hosted on github each of those three points may be provided implicitly.
https://github.com/openjournals/joss-reviews/issues/4104
