### 2022-02-03 neutrons/PyRS#783 @andyfaff
JOSS: it's not clearly described in the manuscript how this software compares to other commonly used packages.
https://github.com/openjournals/joss-reviews/issues/4104

This doesn't apply for the reduction and processing capabilities of the software, as it is bespoke to the instrumentation at ORNL. However, the JOSS manuscript does not describe how PyRS' data analysis capabilities compare to other available packages. The JOSS guidelines suggest the authors should add this.
