### 2022-02-03 #4104 @andyfaff
@rkurchin, I've reviewed the manuscript, and ticked the things that I think meet the JOSS guidelines. I have submitted issues on the [PyRS repository](https://github.com/neutrons/PyRS/) corresponding to the items on the review checklist that don't meet the guideline requirements. The most significant of these is not being able to install the package according to the supplied instructions (and therefore check functional claims, run tests, etc).
