### 2022-02-07 71839b5
Updated manual reduction test
