### 2022-02-07 2e1a2a9
updated pyrs installation and launch in readme
