### 2022-02-07 neutrons/PyRS#778 @fanchercm
The underlying mantid dependencies have evolved. We had not properly vetted the README to ensure that it kept up with the evolution.
```
conda create -n pyrs -c mantid -c mantid/label/nightly mantid-workbench -c conda-forge  --file requirements.txt --file requirements_dev.txt
```

was modified to:
```
conda config --add channels conda-forge --add channels mantid --add channels mantid/label/nightly
conda install mamba
mamba create -n pyrs python=3.6 --file requirements.txt --file requirements_dev.txt
```

launching pyRS was not modified from the pre-mantid 5.0 upgrade:

```
PYTHONPATH=$PWD:$PYTHONPATH QT_API=pyqt python scripts/pyrsplot
```

is now

```
PYTHONPATH=$PWD:$PYTHONPATH python scripts/pyrsplot
```
The mantid conda package is fairly out of date and has paused maintaining builds for macOS and is locked to python 3.6. As such, the operating system and python versions were updated to reflect the current state of the mantid requirement.

We created a branch JOSS_ISSUE with the updates
