+++ README.rst
@@ -1,10 +1,30 @@
+pyRS
+=========
+
 .. image:: https://github.com/neutrons/PyRS/actions/workflows/ci.yml/badge.svg?branch=next
   :target: https://github.com/neutrons/PyRS/actions?query=branch:next
 
 .. image:: https://codecov.io/gh/neutrons/PyRS/branch/master/graph/badge.svg
   :target: https://codecov.io/gh/neutrons/PyRS
 
+Data reduction workflow
 -----------------------
+
+  The workflow is as follows.
+  The word "script" is used to denote a distict step in processing the data rather than actual script.
+
+  1. Start with a nexus file. As in, ``/HFIR/HB2B/IPTS-22731/nexus/HB2B_439.nxs.h5``
+  2. This is read in by something that will split into sub-scans and create a project file [script 1 - currently convert_nexus_to_hidra.py]
+  3. This is read in by something that will create powder patterns and add them to the project file [script 2 - currently reduce_HB2B.py]
+  4. This is read in by the gui to do the peak fitting and add the results to the project file [script 3 - currently somewhere in the gui]
+  5. This is read in by the gui and used to create the summary .csv file [script 4 - currently being written]
+
+  Some other things to note
+  * scripts 1-2 will be refactored/combined to create a new reduce_HB2B.py that can be run by autoreduction
+  * scripts 2-4 (more like functionality 2-4) will all live in the gui in some way
+  * going back any number of steps will delete the following steps from the project file. This is to prevent users from having a project file with powder patterns that are not associated with the peak fitting from the project file.
+
+
 Developer Quick Start
 -----------------------
 
@@ -12,6 +32,10 @@ If you've never used PyRS before, you can get started quickly by doing the follo
 
 1. Install basic dependencies: Conda, Python, and PyQt. Conda installation requires linux.
 2. Create a new Conda environment with additional dependencies:
+
+Installation using an anaconda environment
+------------
+Anaconda environments are only supported on Linux using python 3.6
 
 .. code-block::
    $ conda config --add channels conda-forge --add channels mantid --add channels mantid/label/nightly
@@ -34,32 +58,13 @@ If you've never used PyRS before, you can get started quickly by doing the follo
 
 .. code-block::
 
-   $ pyrsplot
+   $ PYTHONPATH=$PWD:$PYTHONPATH python scripts/pyrsplot
 
 For those setting up in a virtual environment, the dependencies are
 listed in ``requirements.txt`` and ``requirements_dev.txt``.  They can
 be supplide to ``conda create`` using the `--file`` argument, or to
 ``pip install`` using the ``-R`` argument.
 
------------------------
-Data reduction workflow
------------------------
-
-The workflow is as follows.
-The word "script" is used to denote a distict step in processing the data rather than actual script.
-
-1. Start with a nexus file. As in, ``/HFIR/HB2B/IPTS-22731/nexus/HB2B_439.nxs.h5``
-2. This is read in by something that will split into sub-scans and create a project file [script 1 - currently convert_nexus_to_hidra.py]
-3. This is read in by something that will create powder patterns and add them to the project file [script 2 - currently reduce_HB2B.py]
-4. This is read in by the gui to do the peak fitting and add the results to the project file [script 3 - currently somewhere in the gui]
-5. This is read in by the gui and used to create the summary .csv file [script 4 - currently being written]
-
-Some other things to note
-* scripts 1-2 will be refactored/combined to create a new reduce_HB2B.py that can be run by autoreduction
-* scripts 2-4 (more like functionality 2-4) will all live in the gui in some way
[…]
