### 2022-02-07 ba91402
added contributions methods
