+++ README.rst
@@ -37,7 +37,8 @@ Installation using an anaconda environment
 ------------
 Anaconda environments are only supported on Linux using python 3.6
 
-1. Configure anaconda environment:
+1. Configure anaconda environment
+
 .. code-block::
   conda config --add channels conda-forge --add channels mantid --add channels mantid/label/nightly
   conda install mamba
@@ -48,19 +49,19 @@ Anaconda environments are only supported on Linux using python 3.6
 
 .. code-block::
 
-   $ conda activate pyrs
+   conda activate pyrs
 
 3. From the PyRS directory, run the setup script in developer mode
 
 .. code-block::
 
-   $ python setup.py build
+   python setup.py build
 
 4. From the PyRS directory, start the user interface
 
 .. code-block::
 
-   $ PYTHONPATH=$PWD:$PYTHONPATH python scripts/pyrsplot
+   PYTHONPATH=$PWD:$PYTHONPATH python scripts/pyrsplot
 
 For those setting up in a virtual environment, the dependencies are
 listed in ``requirements.txt`` and ``requirements_dev.txt``.  They can
