### 2022-02-07 #4104 @fanchercm
I have started looking at the issues opened by @andyfaff<https://github.com/andyfaff>. I created at branch https://github.com/neutrons/PyRS/tree/JOSS_ISSUE that has corrected details about installing and using the pyrs code.
