### 2022-02-21 neutrons/PyRS#781 @fanchercm
Additional modules were included in the readthedocs to better cover the key functions and classes in the pyrs API
