### 2022-02-21 neutrons/PyRS#779 @fanchercm
New installation instructions in #778 should allow the reviewer to install and use pyRS on a linux machine.
