+++ docs/rtd/source/basics/basics.rst
@@ -31,8 +31,8 @@ Peak Fitting
 Overview of the pyRS peak fitting UI
 ------------------------------------
 
-.. image:: peak_overview.png
-  :width: 600
+.. image:: peak_overview.svg
+  :width: 800
   :alt: pyrsplot startup window
 
 Fitting measured data using pyRS
@@ -41,29 +41,29 @@ Fitting measured data using pyRS
 pyRS was designed to allow the software user to define the peak fitting of data stored using the hidraprojectfile definition. Users interact with the peak fitting UI by defining N ranges for individual peak fitting, where N is the number of windows of interest. Data are loaded by either selecting a specific project file or on the anlaysis cluster by loading a specific run number. A user defines a peak window interactively in the UI by clicking (and holding) on one side of the peak dragging over the specific peak of interest. The user can tweek the graphically defined window by double clicking on the x_left, x_right, or Label entry intry in Peak Ranges. Users can export these inputs as a json file for use in later sessions.
 
 .. image:: define_range.png
-  :width: 600
+  :width: 800
   :alt: pyrsplot startup window
 
 Below are examples of peak fits that use a single an multiple fit windows.
 
 .. image:: single_fit.png
-  :width: 600
+  :width: 800
   :alt: pyrsplot startup window
 
 .. image:: multi_fit.png
-  :width: 600
+  :width: 800
   :alt: pyrsplot startup window
 
 After defining the range click "Fit Peak(s)" to launch the anlaysis.
 
 .. image:: fit_data.png
-  :width: 600
+  :width: 800
   :alt: pyrsplot startup window
 
 Results from the peak fitting are visualized on the right using 1D or 2D scatter plots. Users can define what paramters are visulized by changing the 1D or 3D scatter paramters
 
 .. image:: visualize_res.png
-  :width: 600
+  :width: 800
   :alt: pyrsplot startup window
 
 Stress Strain Analysis
