+++ docs/rtd/source/basics/basics.rst
@@ -31,7 +31,7 @@ Peak Fitting
 Overview of the pyRS peak fitting UI
 ------------------------------------
 
-.. image:: peak_overview.svg
+.. image:: ../figures/peak_overview.svg
   :width: 800
   :alt: pyrsplot startup window
 
@@ -40,29 +40,29 @@ Fitting measured data using pyRS
 
 pyRS was designed to allow the software user to define the peak fitting of data stored using the hidraprojectfile definition. Users interact with the peak fitting UI by defining N ranges for individual peak fitting, where N is the number of windows of interest. Data are loaded by either selecting a specific project file or on the anlaysis cluster by loading a specific run number. A user defines a peak window interactively in the UI by clicking (and holding) on one side of the peak dragging over the specific peak of interest. The user can tweek the graphically defined window by double clicking on the x_left, x_right, or Label entry intry in Peak Ranges. Users can export these inputs as a json file for use in later sessions.
 
-.. image:: define_range.png
+.. image:: ../figures/define_range.png
   :width: 800
   :alt: pyrsplot startup window
 
 Below are examples of peak fits that use a single an multiple fit windows.
 
-.. image:: single_fit.png
+.. image:: ../figures/single_fit.png
   :width: 800
   :alt: pyrsplot startup window
 
-.. image:: multi_fit.png
+.. image:: ../figures/multi_fit.png
   :width: 800
   :alt: pyrsplot startup window
 
 After defining the range click "Fit Peak(s)" to launch the anlaysis.
 
-.. image:: fit_data.png
+.. image:: ../figures/fit_data.png
   :width: 800
   :alt: pyrsplot startup window
 
 Results from the peak fitting are visualized on the right using 1D or 2D scatter plots. Users can define what paramters are visulized by changing the 1D or 3D scatter paramters
 
-.. image:: visualize_res.png
+.. image:: ../figures/visualize_res.png
   :width: 800
   :alt: pyrsplot startup window
 
@@ -71,6 +71,28 @@ Stress Strain Analysis
 
 Select the Stress/Strain Calculation option to launch the stress analysis UI. Reminder overivew of the UI interface:
 
-.. image:: ../basics/stress_overview.svg
+.. image:: ../figures/stress_overview.svg
   :width: 800
   :alt: Stress Analysis overivew
+
+Define the stress condition that pyRS will use to calculate the stresses.
+
+  .. image:: ../figures/Stress_Load.png
+    :width: 800
+    :alt: pyrsplot startup window
+
+  Below are examples of peak fits that use a single an multiple fit windows.
+
+  .. image:: ../figures/Stress_Define_Material.png
+    :width: 800
+    :alt: pyrsplot startup window
+
+  .. image:: ../figures/Stress_Define_d0.png
+    :width: 800
+    :alt: pyrsplot startup window
+
+  After defining the range click "Fit Peak(s)" to launch the anlaysis.
+
+  .. image:: ../figures/Stress_Final.png
+    :width: 800
+    :alt: pyrsplot startup window
