+++ docs/rtd/source/example/example.rst
@@ -16,7 +16,7 @@ Peak Fitting
 
 Select the peak fitting option to launch the peak fitting UI. Reminder overivew of the UI interface:
 
-.. image:: ../basics/peak_overview.svg
+.. image:: ../figures/peak_overview.svg
   :width: 800
   :alt: Peak Fitting overivew
 
@@ -29,21 +29,21 @@ Fitting measured data using pyRS
 
 Graphically define a fit window over the peak of interest. Double click in the x_left (87.0), x_right (92.5), and Label (Al_311). Save the fit range for later use. See below for an example of what the UI should look like before Fitting. After saving the range, click the "Fit Peak(s)" button to start the peak fitting process. After fitting, save the results using either "File/Save" or "File/Save As" to save a new hidraprojectfile. Note that the hidraprojectfiles are not write protected and you can overwrite prior peak fitting results.
 
-.. image:: Example_Fit.png
+.. image:: ../figures/Example_Fit.png
   :width: 800
   :alt: pyrsplot startup window
 
 Below are examples of peak fits that use a single an multiple fit windows.
 
-.. image:: Fit_2646.png
+.. image:: ../figures/Fit_2646.png
   :width: 800
   :alt: Fit of run 2646
 
-.. image:: Fit_2647.png
+.. image:: ../figures/Fit_2647.png
   :width: 800
   :alt: Fit of run 2647
 
-.. image:: Fit_2651.png
+.. image:: ../figures/Fit_2651.png
   :width: 800
   :alt: Fit of run 2651
 
@@ -52,6 +52,28 @@ Stress Analysis
 
 Select the Stress/Strain Calculation option to launch the stress analysis UI. Reminder overivew of the UI interface:
 
-.. image:: ../basics/stress_overview.svg
+.. image:: ../figures/stress_overview.svg
   :width: 800
   :alt: Stress Analysis overivew
+
+Define the stress condition that pyRS will use to calculate the stresses.
+
+  .. image:: ../figures/Stress_Load.png
+    :width: 800
+    :alt: pyrsplot startup window
+
+  Below are examples of peak fits that use a single an multiple fit windows.
+
+  .. image:: ../figures/Stress_Define_Material.png
+    :width: 800
+    :alt: pyrsplot startup window
+
+  .. image:: ../figures/Stress_Define_d0.png
+    :width: 800
+    :alt: pyrsplot startup window
+
+  After defining the range click "Fit Peak(s)" to launch the anlaysis.
+
+  .. image:: ../figures/Stress_Final.png
+    :width: 800
+    :alt: pyrsplot startup window
