+++ docs/rtd/source/basics/basics.rst
@@ -38,61 +38,63 @@ Overview of the pyRS peak fitting UI
 Fitting measured data using pyRS
 --------------------------------
 
-pyRS was designed to allow the software user to define the peak fitting of data stored using the hidraprojectfile definition. Users interact with the peak fitting UI by defining N ranges for individual peak fitting, where N is the number of windows of interest. Data are loaded by either selecting a specific project file or on the anlaysis cluster by loading a specific run number. A user defines a peak window interactively in the UI by clicking (and holding) on one side of the peak dragging over the specific peak of interest. The user can tweek the graphically defined window by double clicking on the x_left, x_right, or Label entry intry in Peak Ranges. Users can export these inputs as a json file for use in later sessions.
+pyRS was designed to allow the software user to define the peak fitting of data stored using the hidraprojectfile definition. Users interact with the peak fitting UI by defining N ranges for individual peak fitting, where N is the number of windows of interest. Data are loaded by either selecting a specific project file or on the analysis cluster by loading a specific run number. A user defines a peak window interactively in the UI by clicking (and holding) on one side of the peak dragging over the specific peak of interest. The user can tweak the graphically defined window by double clicking on the x_left, x_right, or Label entry in Peak Ranges. Users can export these inputs as a json file for use in later sessions.
 
 .. image:: ../figures/define_range.png
   :width: 800
-  :alt: pyrsplot startup window
+  :alt: define range
 
-Below are examples of peak fits that use a single an multiple fit windows.
+Below are examples of peak fits that use a single and multiple fit windows.
 
 .. image:: ../figures/single_fit.png
   :width: 800
-  :alt: pyrsplot startup window
+  :alt: single fit example
 
 .. image:: ../figures/multi_fit.png
   :width: 800
-  :alt: pyrsplot startup window
+  :alt: multifit example
 
-After defining the range click "Fit Peak(s)" to launch the anlaysis.
+After defining the range click "Fit Peak(s)" to launch the analysis.
 
 .. image:: ../figures/fit_data.png
   :width: 800
-  :alt: pyrsplot startup window
+  :alt: peak fitting
 
-Results from the peak fitting are visualized on the right using 1D or 2D scatter plots. Users can define what paramters are visulized by changing the 1D or 3D scatter paramters
+Results from the peak fitting are visualized on the right using 1D or 2D scatter plots. Users can define what parameters are visualized by changing the 1D or 3D scatter parameters
 
 .. image:: ../figures/visualize_res.png
   :width: 800
-  :alt: pyrsplot startup window
+  :alt: vis peak fit
 
 Stress Strain Analysis
 ======================
 
-Select the Stress/Strain Calculation option to launch the stress analysis UI. Reminder overivew of the UI interface:
+Select the Stress/Strain Calculation option to launch the stress analysis UI. Reminder overview of the UI interface:
 
 .. image:: ../figures/stress_overview.svg
   :width: 800
   :alt: Stress Analysis overivew
 
-Define the stress condition that pyRS will use to calculate the stresses.
+Define the stress condition that pyRS will use to calculate the stresses. A 2D condition allows the selection of either a plane strain or plane stress condition. A 2D condition only prompts loading 2 hidraprojectfiles with peak fit results. Otherwise, a 3D stress condition requires 3 hidraprojectfiles.
 
   .. image:: ../figures/Stress_Load.png
     :width: 800
-    :alt: pyrsplot startup window
+    :alt: load project files
 
-  Below are examples of peak fits that use a single an multiple fit windows.
+After loading 2, or 3, hidraprojectfiles the users need to enter the Mechanical constants (or Materials parameters) Youngs Modulus and Poisson's ratio.
 
[…]
