### 2022-02-21 neutrons/PyRS#780 @fanchercm
An example has been added to the JOSS_ISSUE with data under the example folder. We have included information in an Example section in the readthedocs
