### 2022-02-22 b0200cb
added JOSS paper to repo
