+++ paper/paper.md
@@ -1,0 +1,104 @@
+---
+title: 'pyRS: A Python package for the reduction and analysis of neutron residual stress data'
+tags:
+  - Python
+  - Mechanics of Materials
+  - Neutron Diffraction
+  - Residual Stress
+authors:
+  - name: Chris M. Fancher # note this makes a footnote saying 'co-first author'
+    orcid: 0000-0002-3952-5168
+    affiliation: "1, 2" # (Multiple affiliations must be quoted)
+  - name: Jean Bilheux
+    orcid: 0000-0003-2172-6487
+    affiliation: 2
+  - name: Wenduo Zhou # note this makes a footnote saying 'co-first author'
+    orcid: 0000-0002-2848-5324
+    affiliation: "3"
+  - name: Ross E. Whitfield # note this makes a footnote saying 'co-first author'
+    orcid: 0000-0002-9852-1044
+    affiliation: "3"
+  - name: Jose Borreguero # note this makes a footnote saying 'co-first author'
+    orcid: 0000-0002-0866-8158
+    affiliation: "3"
+  - name: William F. Godoy # note this makes a footnote saying 'co-first author'
+    orcid: 0000-0002-2590-5178
+    affiliation: "3"
+  - name: Steven Hahn # note this makes a footnote saying 'co-first author'
+    orcid: 0000-0002-2018-7904
+    affiliation: "3"
+  - name: Peter F. Peterson # note this makes a footnote saying 'co-first author'
+    orcid: 0000-0002-1353-0348
+    affiliation: "3"
+affiliations:
+ - name: Materials Science & Technology Division, Oak Ridge National Laboratory, Oak Ridge, TN
+   index: 1
+ - name: Neutron Scattering Division, Oak Ridge National Laboratory, Oak Ridge, TN
+   index: 2
+ - name: Computer Science and Mathematics Division, Oak Ridge National Laboratory, Oak Ridge, TN
+   index: 3
+date: 01 November 2021
+bibliography: paper.bib
+---
+
+# Statement of need
+
+The 2nd Generation Neutron Residual Stress Facility (NRSF2) residual stress mapping instrument at the High Flux Isotope Reactor (HFIR) at ORNL was recently rebuilt with a modern detector and control system. Upgrading from a LabVIEW-based control software (SPICE) to an Experimental Physics and Industrial Control System (EPICS) based control software with the neutron Event Distributor (nED) data acquisition system brought additional experimental flexibility.[@White2019; @Vodopivec2017] The transition from a control system that measured discrete data to an event-based data structure.[@PETERSON201524] deprecated the data reduction and analysis software. The design of `pyRS` relied on years of experience with the previous reduction and analysis software (NRSF View) to ensure the new software builds upon the strengths and improves on the weakness of NRSF View. The lack of a unified analysis and visualization of residual stress data was among the most significant needs.  
+
+# Summary
+`pyRS` is a python software package that was designed to meet the data reduction and analysis needs of the neutron residual stress mapping user community at Oak Ridge National Laboratory (ORNL). `pyRS` incorporates separate modules that provide a streamlined workflow for reducing raw neutron events into 1D intensity vs. scattering angle and subsequent analysis to extract the interatomic spacing and intensity for residual stress and texture analysis. Users can access the modules through either a graphical or command-line interface. `pyRS` saves data into a single HDF5 file.[@hdf5], in which the metadata, reduced diffraction data, and peak analysis results are passed between different modules.
+
+# Overview of pyRS
[…]
