### 2022-02-22 neutrons/PyRS#783 @fanchercm
A discussion of how the pyRS capabilities compare with the community was added to the paper:
The overall functionality of pyRS is similar to the stress texture calculator package developed at FRM-II, which includes modules for data reduction, analysis, and visualization.[@Randau2011] Nuance differences in how raw data are saved necessitated the development of a new package tailored to the needs of the HIDRA instrument and its unique user community. The peak fitting in pyRS utilizes the mantid framework, a standardized collection of algorithms that are tailored for processing neutron diffraction and scattering data.[@Arnold2014a] Mantid implements peak shape functions common to neutron diffraction data (e.g., Gaussian, Lorentz, pseudo-Voigt, Voigt, etc...). The underlying peak fitting capabilities in Mantid are similar to the LMFIT python package.[@Newville2021] The interpretation of peak fitting results using equations 1-3 follow principle stress/strain analysis used by the neutron engineering diffraction community.[@Huchings] While the underlying mathematics are common to the neutron community, the visualization capabilities of pyRS are unique and were designed to facilitate users to visualize neutron mapping results more readily.
