### 2022-02-23 #4104 @andyfaff
@fanchercm, you've been making headway on the review issues I raised. Are the changes solely being made to the [JOSS_ISSUES branch](https://github.com/neutrons/PyRS/commits/JOSS_ISSUE), or are they spread over different branches?

@rkurchin, I have a question for JOSS in general. Do the changes being made to a project have to have been a) committed to the 'main' branch, or in a release (tag), for the JOSS paper to be accepted? Here, the issues are being addressed in a feature branch for the paper, but they've not been merged into the main branch yet.
