### 2022-02-27 neutrons/PyRS#785 @sangpham171
JOSS: Installation: Does installation proceed as outlined in the documentation?
Hello,
I try to run the pyplot, but no success. Is it an obligation to use Linux? because i am using window and python version 3.9.10
I have already installed all the requirements except the mantide-workbench, I can't find with pip.

Here is the problem:

C:\PyRS-next>py scripts/pyrsplot
C:\PyRS-next\scripts\pyrsplot:3: DeprecationWarning: the imp module is deprecated in favour of importlib; see the module's documentation for alternative uses
  import imp
Traceback (most recent call last):
  File "C:\PyRS-next\scripts\pyrsplot", line 6, in <module>
    import pyrs.interface.pyrs_main
ModuleNotFoundError: No module named 'pyrs.interface'

If not, do you have an .exe file to test the program. Because for a normal user it might be difficult to install all the dependencies.
https://github.com/openjournals/joss-reviews/issues/4104
