### 2022-02-28 neutrons/PyRS#785 @fanchercm
We have never tested pyRS from windows. My best guess is that there is a path issue related to running pyrsplot. Please see the JOSS_ISSUE branch for an updated readme. At this moment we do not have compiled version of the software. We are waiting for mantid dependency issues to be resolved before deploying pyRS as part of an anaconda release.
