### 2022-03-01 neutrons/PyRS#785 @fanchercm
@sangpham171  Please try the following:
From the PyRS directory, start the user interface
`set PYTHONPATH=%PATH%;%PYTHONPATH%; python scripts/pyrsplot`
