### 2022-04-01 #4104 @fanchercm
I’m setting up a windows VM to troubleshoot using PyRS on a windows machine. Another option is to see if Stuart Campbell (stuartcampbell) is available for the review.
