### 2022-04-08 #4104 @andyfaff
@rkurchin I'm just coming back to this now, sorry for the delay.
I note that there is already an article on pyRS, https://doi.org/10.1107/S1600576721010554, describing what the software does. There is [guidance for this](https://joss.readthedocs.io/en/latest/submitting.html#co-publication-of-science-methods-and-software). AFAIK the authors didn't indicate "whether related publications (published, in review, or nearing submission) exist as part of submitting to JOSS". It's not clear as a reviewer what I need to do with regards to that, so decisions/input on that are in the editors court.
