### 2022-04-08 neutrons/PyRS#778 @andyfaff
The install instructions on the JOSS_ISSUE branch now get me to a state where I can get going with running the GUI.
However, the Python version referred to in those instructions is 3.6, which is past EOL. I gather that this restriction is due to Mantid itself, but in my opinion it would look funny publishing a JOSS paper for a Python version which has already EOLd.
