### 2022-04-08 neutrons/PyRS#786 @andyfaff
JOSS: running the tests.
On the JOSS_ISSUE branch the test suite has 5 failures. These are:

Several of them appear with a file dialog to manually select files. I have no idea of which files to select.

(xref https://github.com/openjournals/joss-reviews/issues/4104)
