### 2022-04-08 neutrons/PyRS#788 @andyfaff
JOSS: can't find the example files
I'm trying to run the examples at https://neutrons-pyrs.readthedocs.io/en/latest/example/example.html#pyrs-example-use, but can't find the examples referred to [here](https://neutrons-pyrs.readthedocs.io/en/latest/example/example.html#load-measured-data). In which directory are they?

(xref https://github.com/openjournals/joss-reviews/issues/4104)
