### 2022-04-08 neutrons/PyRS#788 @fanchercm
From the JOSS_ISSUE branch the files are under the "example" folder. I have updated the readthedocs to help clarify the location of the data.

'example/HB2B_2246.h5'
'example/HB2B_2247.h5'
'example/HB2B_2251.h5'
