+++ pyrs/calibration/peakfit_calibration.py
@@ -177,12 +177,12 @@ class PeakFitCalibration:
             self.engines.append([powder_engine, powder_lines, single_material])
 
         # calibration: numpy array. size as 7 for ... [6] for wave length
-        self._calib = np.array(8 * [0], dtype=np.float)
+        self._calib = np.array(8 * [0], dtype=np.float64)
         # calibration error: numpy array. size as 7 for ...
-        self._caliberr = np.array(8 * [-1], dtype=np.float)
+        self._caliberr = np.array(8 * [-1], dtype=np.float64)
 
         # calibration starting point: numpy array. size as 7 for ...
-        self._calib_start = np.array(8 * [0], dtype=np.float)
+        self._calib_start = np.array(8 * [0], dtype=np.float64)
 
         if wavelength is None:
             # Set wave length
@@ -658,7 +658,7 @@ class PeakFitCalibration:
 
         paramVec = np.copy(self._calib)
         paramVec[i_index] = x
-        print(x)
+
         residual = self.get_alignment_residual(paramVec, roi_vec_set, ConstrainPosition, False, start, stop)
 
         if ReturnScalar:
@@ -676,7 +676,7 @@ class PeakFitCalibration:
         """
         paramVec = np.copy(self._calib)
         paramVec[0:3] = x[:]
-        print(x)
+
         residual = self.get_alignment_residual(paramVec, roi_vec_set, ConstrainPosition, False, start, stop)
 
         if ReturnScalar:
