+++ environment.yml
@@ -1,5 +1,6 @@
 name: PyRS
 dependencies:
+- python=3.8 
 - mantidworkbench
 - qtpy
 - pytest
