### 2022-04-09 59355da
updated enviroment to mantid nightly
