+++ environment.yml
@@ -1,8 +1,7 @@
 name: PyRS
 dependencies:
-- python=3.8 
-- mantid
-- mantidqt
+- python=3.8
+- mantidworkbench
 - qtpy
 - pytest
 - pytest-cov
