### 2022-04-09 neutrons/PyRS#786 @fanchercm
Some screen resolutions can cause an error to occur. The tests have been updated in JOSS_ISSUE to address the test failures.
