+++ pyrs/peaks/peak_collection.py
@@ -272,7 +272,7 @@ class PeakCollection:
         return self._sub_run_array
 
     @property
-    def fitting_costs(self) -> np.ndarray:
+    def fitting_costs(self):
         return self._fit_cost_array
 
     @property
@@ -436,7 +436,7 @@ class PeakCollection:
         d_spacing = self._get_dspacing_center()
         return unumpy.nominal_values(d_spacing), unumpy.std_devs(d_spacing)
 
-    def get_chisq(self) -> np.ndarray:
+    def get_chisq(self):
         return np.copy(self._fit_cost_array)
 
     def get_subruns(self):
