### 2022-04-11 cb7bd3e
Added check for new mantid that changes paths
