### 2022-04-11 #4104 @rkurchin
@fanchercm, apologies, yes I can see that now. Quoting from the guidelines:

> ### [Co-publication of science, methods, and software](https://joss.readthedocs.io/en/latest/submitting.html#co-publication-of-science-methods-and-software)
> Sometimes authors prepare a JOSS publication alongside a contribution describing a science application, details of algorithm development, and/or methods assessment. In this circumstance, JOSS considers submissions for which the implementation of the software itself reflects a substantial scientific effort. This may be represented by the design of the software, the implementation of the algorithms, creation of tutorials, or any other aspect of the software. We ask that authors indicate whether related publications (published, in review, or nearing submission) exist as part of submitting to JOSS.

As the person by far the most familiar with both pieces of work, could you clarify what the differing focuses are of the two papers such that these guidelines are adhered to?
