### 2022-04-11 #4104 @fanchercm
The scientific paper was published in the journal of applied crystallography and details the specifics of the science application and crystallography as well as specifics of the data handling needed to ensure neutrons are properly reduced using established procedures. The JOSS paper was initially geared toward the software development. Looking at the guidelines, I think it would be good to include the tutorial created as part of issue #780 in the the paper.
