### 2022-04-13 c241559
fixed gui errors with newer dependencies
