+++ pyrs/interface/ui/mplfitplottingwidget.py
@@ -418,7 +418,7 @@ class QtMplFitCanvas(FigureCanvas):
         vec_y = data_set[1]
 
         # print ('[DB...BAT] Plot residual:\n{}\n{}'.format(vec_x, vec_y))
-
+        self._residual_subplot.cla()
         plot_info = self._residual_subplot.plot(vec_x, vec_y, label=None, color='green',
                                                 linestyle='-', linewidth=2)
         self._residual_subplot.axhline(2, linestyle='--', color='blue')
@@ -463,6 +463,7 @@ class QtMplFitCanvas(FigureCanvas):
                                             linestyle=line_style, linewidth=line_width)
 
         self._data_subplot.set_aspect('auto')
+        self._data_subplot.set_xticklabels([])
 
         real_max = np.nanmax(vec_y)
         max_value_with_offset = real_max + 0.1*real_max
