+++ environment.yml
@@ -1,7 +1,7 @@
 name: PyRS
 dependencies:
 - python=3.8
-- mantidworkbench
+- mantid/label/nightly::mantidworkbench
 - qtpy
 - pytest
 - pytest-cov
