### 2022-04-13 #4104 @arfon
:wave: all, EiC of JOSS here.

> The scientific paper was published in the journal of applied crystallography and details the specifics of the science application and crystallography as well as specifics of the data handling needed to ensure neutrons are properly reduced using established procedures.

@fanchercm – could you say more about how the state of the software as published in https://doi.org/10.1107/S1600576721010554 differs from what is being reviewed here? Have there been significant updates to the software since https://doi.org/10.1107/S1600576721010554 was reviewed and published?
