### 2022-04-25 #4104 @arfon
Thanks for the additional information @fanchercm. Personally, I'm still unclear whether these changes warrant a new publication and meet our [substantial scholarly effort](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort) criterion.

Could you please provide a link in your commit history that represents the state of the project at the time of the earlier publication so we can compare to the current state? https://github.com/neutrons/PyRS/commits/next
