### 2022-05-12 #4104 @fanchercm
Thank you for the quick response!

We had initially structured the two papers based on the IGRINS RV work (https://joss.theoj.org/papers/10.21105/joss.03095 and https://iopscience.iop.org/article/10.3847/1538-3881/abf5e7/pdf). Our team is not familiar with astronomical papers. We might have missinterupted the differences between the two sets of papers.

Chrsi
