+++ pyrs/interface/manual_reduction/event_handler.py
@@ -1,11 +1,11 @@
 import os
-from qtpy.QtCore import Qt
-from qtpy.QtWidgets import QApplication
+from qtpy.QtCore import Qt # type: ignore
+from qtpy.QtWidgets import QApplication # type: ignore
 from pyrs.interface.gui_helper import pop_message, browse_file, browse_dir, parse_combo_box
 from mantidqt.utils.asynchronous import BlockingAsyncTaskWithCallback
 from pyrs.interface.manual_reduction.pyrs_api import ReductionController
-from pyrs.dataobjects.constants import HidraConstants  # type: ignore
-from pyrs.utilities import get_default_output_dir, get_nexus_file  # type: ignore
+from pyrs.dataobjects.constants import HidraConstants # type: ignore
+from pyrs.utilities import get_default_output_dir, get_nexus_file # type: ignore
 
 
 class EventHandler:
