### 2022-05-25 75c28e9
Change how result is logged
