+++ .github/workflows/ornl-prod.yml
@@ -1,10 +1,0 @@
-name: Deploy to ORNL
-on:
-  push:
-    branches:
-      - master
-jobs:
-  ornl-trigger:
-    runs-on: ubuntu-20.04
-    steps:
-      - run: curl -sX POST -F token=${{ secrets.GITLAB_TRIGGER_TOKEN }} -F ref=master ${{ secrets.GITLAB_TRIGGER_URL }} | jq '. | del( .user )'
