### 2022-05-25 f2dd6c8
Use newer curl and fix variable name
