### 2022-05-25 5970077
Hide jq call for debugging purposes
