### 2022-05-25 82db840
Add more diagnostics and get non-zero return
