### 2022-05-26 f1b27d2
Intentionally failing to try using pipestatus
