### 2022-06-04 #4104 @arfon
Hi @fanchercm – thanks for the links and responses and I'm sorry there have been communication issues on this submission.

> We had initially structured the two papers based on the IGRINS RV work (https://joss.theoj.org/papers/10.21105/joss.03095 and https://iopscience.iop.org/article/10.3847/1538-3881/abf5e7/pdf). Our team is not familiar with astronomical papers. We might have missinterupted the differences between the two sets of papers.

Thanks is a helpful clarification and helps me understand what is going on here.

As you have noted, our [partnership with AAS publishing](https://blog.joss.theoj.org/2018/12/a-new-collaboration-with-aas-publishing) allows for two papers to be associated with essentially the same work. The AAS paper generally focused on the methods, scientific applicability (including possibly novel results), and the JOSS paper focused specifically on the software (and the review of it).

Unfortunately we don't offer this as a general solution for all authors/all disciplines. JOSS papers published with AAS are clearly marked as joint papers, citing each other, and linking between in the JOSS PDF and on the AAS publishing website.

While I recognize what you're attempting to do here (i.e., replicate the JOSS/AAS process), we would consider this paper in JOSS a duplicate (i.e., not allowed) of the [Journal of Applied Crystallography paper](https://doi.org/10.1107/S1600576721010554) you already published.

My [question above](https://github.com/openjournals/joss-reviews/issues/4104#issuecomment-1098253222) about the changes made to the software since the Journal of Applied Crystallography paper we're designed to see if the changes since that paper might warrant a _second_ (i.e., non-duplicate) paper about this work. Unfortunately I don't think [these changes](https://github.com/neutrons/PyRS/compare/v1.1.0...next) meet our [Substantial Scholarly Effort](https://blog.joss.theoj.org/2020/07/minimum-publishable-unit) criterion.

As such, with regret, I'm afraid we must reject this submission.
