### 2022-10-11 #4104 @fanchercm
Hi @rkurchin our mail system has implemented new security. I’m not sure if they are blocking notification.

***@***.***  and @arfon, I’m more confused about what the criteria is for the “substantial scholarly effort”. I was under the impression that some JOSS papers were published as companions to papers detailing the science application. JOSS states “In this circumstance, JOSS considers submissions for which the implementation of the software itself reflects a substantial scientific effort.” The JOSS paper was submitted while the Journal of Applied Crystallography paper was still in Review. The Journal of Applied Crystallography was meant to serve as a high level overview for the science application.

For your review, the V1.1.0 codebase was used for the paper https://github.com/neutrons/PyRS/compare/v1.1.0...next”
