### 2022-10-11 #4104 @fanchercm
Hi EiC,

The referred JAC publication provided a proof-of-concept framework for the pyRS methodology and workflow and presented specific details of the physics. The software capabilities and quality of the same was not peer-reviewed at that time. We did not think the software was in a sufficient state (e.g., documentation, tutorials, examples, ….) to meet JOSS peer-review standards. These deficiencies have been addressed with a fully developed capabilities and document with a readthedoc containing a demonstration example for how a user would analyze independent measurements to determine the final stress state. Subsequent evolution of the mantid dependency have forced further software development, such as a redevelopment of the stress/strain UI backend that is critical for users to visualize results. Evolution of the mantid dependency allowed the development of a secondary reduction pipeline that has enabled read-time visualization of data that has improved instrument productivity. The software efforts are critical for keeping pyRS up to date with Mantid most recent releases, documentation, bugs, and overall software quality for the users. Recent evolution of the legacy dependencies enables PyRS to utilize python 3.8 instead of the EOL python 3.6 that was required for version in the JAC paper.  The JOSS publication seeks to give credit to those involved in the software process to reach acceptable quality levels.
